[#] Rename project and main package.
