+++ pytest_testrail_integrator/dto.py
@@ -5,7 +5,7 @@ from _pytest._code import ExceptionInfo
 from _pytest._code.code import TerminalRepr
 from _pytest.mark import Mark
 
-from pytest_testrail_client.constants import PytestStatus
+from pytest_testrail_integrator.constants import PytestStatus
 
 
 @dataclasses.dataclass
